refactor(vna_calibration): promote out of live_status
